hf-space/requirements.txt
torch
torchvision
transformers
Pillow
gradio
spaces
loadimg
